fixes #7 - see long desc for why
It turns out that this was an issue with redis-server 2.4.14, which
didn't like how the separate SET and EX commands were coming through.
It seems 2.6.16 works fine on the other hand.

I've set the command to SETEX as per @jsimnz's suggestion, which works
fine on both server versions.

I came across this myself while setting up a debian wheezy server.