MaxAge = 0 does not agree wtih SETEX
A cookie MaxAge and the TTL for SETEX aren't necessarily mutal.  A MaxAge
of 0 tells the browser it is a session, but SETEX will fail.  Instead
detect if the MaxAge value is 0 and default to something reasonable for
Redis.  This is currently 20 minutes, and can be set through the
DefaultMaxAge member.